One class per file # Description - Split data.ts into separate files - Moved UI classes into ui-components directory - Updated chart base classes
