+++ src/models/custom-dataset.ts
@@ -1,0 +1,13 @@
+export class CustomDatasetInfo {
+  id: number;
+  name: string;
+  xData: string[];
+  yData: string[];
+
+  constructor() {
+    this.id = -1;
+    this.name = '';
+    this.xData = [];
+    this.yData = [];
+  }
+}
