+++ src/models/margin.ts
@@ -1,0 +1,13 @@
+export class Margin {
+  top: number;
+  right: number;
+  bottom: number;
+  left: number;
+
+  constructor(top: number, right: number, bottom: number, left: number) {
+    this.top = top;
+    this.right = right;
+    this.bottom = bottom;
+    this.left = left;
+  }
+}
