+++ src/models/size.ts
@@ -1,0 +1,9 @@
+export class Size {
+  width: number;
+  height: number;
+
+  constructor(w: number, h: number) {
+    this.width = w;
+    this.height = h;
+  }
+}
